refactor for reuses (WIP)
