+++ src/implementation/x86/mod.rs
@@ -5,7 +5,7 @@ pub(crate) mod avx2;
 pub(crate) mod sse42;
 
 #[allow(unused_imports)]
-use super::SIMD_CHUNK_SIZE;
+use super::algorithm::SIMD_CHUNK_SIZE;
 
 // validate_utf8_basic() std: implementation auto-selection
